clippy: fix 9 latent lints in test-mode code paths
The always-on clippy aspect (added last commit round) clippy-checks
host `bazel build`, but `#[cfg(test)]` code only compiles under
`bazel test` — and that compile was never clippy-clean. The first
`bazel test //crates/...` after the gate landed surfaced 9 lints,
all in test fixtures / test-gated modules:

  * `net/tcp/src/tests.rs`, `quic/src/conn/cfg.rs` — inner
    `#![cfg(test)]` in a module whose `mod` include is already
    `#[cfg(test)]` (clippy::duplicated_attributes).
  * `quic/src/crypto.rs` ×2 lines — `nyb(_) as u8` where `nyb`
    already returns `u8` (clippy::unnecessary_cast).
  * `quic/src/wire.rs` — `101 as u8` literal cast; and a
    `for i in 0..n { out[i] }` index loop (needless_range_loop).
  * `net/src/dhcp_parse.rs` — `val.len() >= 1` → `!is_empty()`
    (clippy::len_zero).
  * `iobuf/src/pool.rs` — `let buf = …; buf` tail (let_and_return).
  * `iobuf/src/iobuf.rs` ×2 — `(b'a'..=b'z').contains(b)` →
    `b.is_ascii_lowercase()` (manual_is_ascii_check).

All fixes are behaviour-preserving. `bazel test //crates/...` now
passes 28/28 with the clippy aspect active.

+++ crates/proto/quic/src/crypto.rs
@@ -306,7 +306,7 @@ mod tests {
 
     // ── tiny hex parsers (const) ─────────────────────────────────
     const fn parse_hex_byte(c0: u8, c1: u8) -> u8 {
-        ((nyb(c0) as u8) << 4) | (nyb(c1) as u8)
+        (nyb(c0) << 4) | nyb(c1)
     }
     const fn nyb(c: u8) -> u8 {
         match c {
@@ -520,7 +520,7 @@ mod tests {
             let mut i = 0;
             let mut j = 0;
             while i < hex_bytes.len() {
-                out[j] = ((nyb_for(hex_bytes[i]) as u8) << 4) | nyb_for(hex_bytes[i + 1]) as u8;
+                out[j] = (nyb_for(hex_bytes[i]) << 4) | nyb_for(hex_bytes[i + 1]);
                 i += 2;
                 j += 1;
             }
